selftests/bpf: Fix incorrect TRUNNER_BINARY name output
Currently, when we run 'make test_progs', the output is:

  CLNG-BPF [test_maps] atomic_bounds.o
  ...
  GEN-SKEL [test_progs] atomic_bounds.skel.h
  ...
  TEST-OBJ [test_progs] align.test.o
  ...
  TEST-HDR [test_progs] tests.h
  EXT-OBJ  [test_progs] test_progs.o
  ...
  BINARY   test_progs

As you can see, the TRUNNER_BINARY name in the CLNG-BPF part is test_maps,
which is incorrect.

Similarly, when we run 'make test_maps', the output is:

  CLNG-BPF [test_maps] atomic_bounds.o
  ...
  GEN-SKEL [test_progs] atomic_bounds.skel.h
  ...
  TEST-OBJ [test_maps] array_map_batch_ops.test.o
  ...
  TEST-HDR [test_maps] tests.h
  EXT-OBJ  [test_maps] test_maps.o
  ...
  BINARY   test_maps

At this time, the TRUNNER_BINARY name in the GEN-SKEL part is wrong.

Again, if we run 'make /full/path/to/selftests/bpf/test_vmlinux.skel.h',
the output is:

  CLNG-BPF [test_maps] test_vmlinux.o
  GEN-SKEL [test_progs] test_vmlinux.skel.h

Here, the TRUNNER_BINARY names are inappropriate and meaningless, they
should be removed.

This patch fixes these and all other similar issues.

With the patch applied, the output becomes:

  $ make test_progs

  CLNG-BPF [test_progs] atomic_bounds.o
  ...
  GEN-SKEL [test_progs] atomic_bounds.skel.h
  ...
  TEST-OBJ [test_progs] align.test.o
  ...
  TEST-HDR [test_progs] tests.h
  EXT-OBJ  [test_progs] test_progs.o
  ...
  BINARY   test_progs

  $ make test_maps

  CLNG-BPF [test_maps] atomic_bounds.o
  ...
  GEN-SKEL [test_maps] atomic_bounds.skel.h
  ...
  TEST-OBJ [test_maps] array_map_batch_ops.test.o
  ...
  TEST-HDR [test_maps] tests.h
  EXT-OBJ  [test_maps] test_maps.o
  ...
  BINARY   test_maps

  $ make /full/path/to/selftests/bpf/test_vmlinux.skel.h

  CLNG-BPF test_vmlinux.o
  GEN-SKEL test_vmlinux.skel.h

Signed-off-by: Yuntao Wang <ytcoode@gmail.com>