crash_core: Fix the check for whether crashkernel is from high memory
The purpose of the reserve_crashkernel_generic() function is to allocate a
block of memory for crash kernel, and if the block of memory is allocated
from high memory, it will allocate an additional block from low memory.

The method to determine if a block of memory is from high memory is to
check if crash_base is greater than or equal to CRASH_ADDR_LOW_MAX.
However, the current code only considers the case where crash_base is
greater than CRASH_ADDR_LOW_MAX.

This means that if the memory is allocated from high memory and its
starting address is CRASH_ADDR_LOW_MAX, reserve_crashkernel_generic()
will no longer allocate the additional memory from low memory for crash
kernel, even if it is necessary.

In fact, we can also take a look at the code before it was modified in
these commits:

  9c08a2a139fe ("x86: kdump: use generic interface to simplify crashkernel reservation code")
  fdc268232dbb ("arm64: kdump: use generic interface to simplify crashkernel reservation")
  39365395046f ("riscv: kdump: use generic interface to simplify crashkernel reservation")

They all checked for the case where crash_base is equal to
CRASH_ADDR_LOW_MAX.

This patch also includes some minor cleanups.

Fixes: 0ab97169aa05 ("crash_core: add generic function to do reservation")
Signed-off-by: Yuntao Wang <ytcoode@gmail.com>