mm/zswap: change dstmem size to one page
Change the dstmem size from 2 * PAGE_SIZE to only one page since we only
need at most one page when compress, and the "dlen" is also PAGE_SIZE in
acomp_request_set_params().  If the output size > PAGE_SIZE we don't wanna
store the output in zswap anyway.

So change it to one page, and delete the stale comment.

Link: https://lkml.kernel.org/r/20231213-zswap-dstmem-v1-2-896763369d04@bytedance.com
Signed-off-by: Chengming Zhou <zhouchengming@bytedance.com>
Reviewed-by: Yosry Ahmed <yosryahmed@google.com>
Cc: Chris Li <chriscli@google.com>
Cc: Dan Streetman <ddstreet@ieee.org>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Nhat Pham <nphamcs@gmail.com>
Cc: Seth Jennings <sjenning@redhat.com>
Cc: Vitaly Wool <vitaly.wool@konsulko.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>