x86/efistub: fix the missing KASLR_FLAG bit in boot_params->hdr.loadflags
When KASLR is enabled, the KASLR_FLAG bit in boot_params->hdr.loadflags
should be set to 1 to propagate KASLR status from compressed kernel to
kernel, just as the choose_random_location() function does.

Currently, when kernel is booted via efi stub, the KASLR_FLAG bit in
boot_params->hdr.loadflags is not set, even though it should be. This
causes some functions, such as kernel_randomize_memory(), not to execute
as expected. Fix it.

Signed-off-by: Yuntao Wang <ytcoode@gmail.com>