percpu: simplify the logic of pcpu_alloc_first_chunk()
In the logic for hiding the end region of the chunk, there are several
places where the same value 'region_bits - offset_bits' is calculated
over and over again using different methods. Eliminating these redundant
calculations can improve code readability.

Additionally, there is a lot of repetitive code when hiding the start and
end regions of the chunk. We can consolidate this logic into a function,
making the final code cleaner and more concise.

Signed-off-by: Yuntao Wang <ytcoode@gmail.com>